test(test-optimization): exercise Vitest EFD helper in CI
